+++ packages/calcite-components/src/components/shell-panel/interfaces.ts
@@ -1,1 +1,10 @@
 export type DisplayMode = "dock" | "float" | "overlay" | "float-content" | "float-all";
+
+export interface ResizeValues {
+  inlineSize: number;
+  blockSize: number;
+  minInlineSize: number;
+  minBlockSize: number;
+  maxInlineSize: number;
+  maxBlockSize: number;
+}