+++ packages/calcite-components/src/custom-theme/time-picker.ts
@@ -1,0 +1,14 @@
+import { html } from "../../support/formatting";
+
+export const timePickerTokens = {
+  calciteTimePickerBackgroundColor: "",
+  calciteTimePickerCornerRadius: "",
+  calciteTimePickerButtonBackgroundColorHover: "",
+  calciteTimePickerButtonBackgroundColorPress: "",
+  calciteTimePickerColor: "",
+  calciteTimePickerIconColor: "",
+  calciteTimePickerInputBorderColorPress: "",
+  calciteTimePickerInputBorderColorHover: "",
+};
+
+export const timePicker = html`<calcite-time-picker></calcite-time-picker>`;