+++ packages/calcite-tailwind-preset/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.0-next.6](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-tailwind-preset@0.2.0-next.5...@esri/calcite-tailwind-preset@0.2.0-next.6) (2025-04-29)
+
+**Note:** Version bump only for package @esri/calcite-tailwind-preset
+
+## [0.2.0-next.5](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-tailwind-preset@0.2.0-next.4...@esri/calcite-tailwind-preset@0.2.0-next.5) (2025-04-23)
+
+**Note:** Version bump only for package @esri/calcite-tailwind-preset
+
+## [0.2.0-next.4](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-tailwind-preset@0.2.0-next.3...@esri/calcite-tailwind-preset@0.2.0-next.4) (2025-04-22)
+
+**Note:** Version bump only for package @esri/calcite-tailwind-preset
+
+## [0.2.0-next.3](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-tailwind-preset@0.2.0-next.2...@esri/calcite-tailwind-preset@0.2.0-next.3) (2025-04-18)
+
+**Note:** Version bump only for package @esri/calcite-tailwind-preset
+
+## [0.2.0-next.2](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-tailwind-preset@0.2.0-next.1...@esri/calcite-tailwind-preset@0.2.0-next.2) (2025-04-18)
+
+**Note:** Version bump only for package @esri/calcite-tailwind-preset
+
+## [0.2.0-next.1](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-tailwind-preset@0.2.0-next.0...@esri/calcite-tailwind-preset@0.2.0-next.1) (2025-04-18)
+
+**Note:** Version bump only for package @esri/calcite-tailwind-preset
+
 ## 0.2.0-next.0 (2025-04-09)
 
 ### Features